requirements.txt
elitegamez
==========

Elitegamez Django Project

Django==1.6.5
MySQL-python==1.2.5
Pillow==2.5.3
South==1.0
admin-multiupload==0.1
boto==2.32.1
django-admin-tools==0.5.2
django-bootstrap3==4.11.0
django-boto==0.3.1
django-carton==1.2.0
django-countries==2.1.2
django-crispy-forms==1.4.0
django-jfu==2.0.9
django-mathfilters==0.3.0
django-model-utils==2.2
django-money==0.4.2
django-money-rates==0.3.1
django-photologue==2.8.3
django-sortedm2m==0.7.0
django-storages==1.1.8
py-moneyed==0.5.0
pytz==2014.7
requests==2.3.0
six==1.7.3
usethis-django-bootstrap==3.2.1
